refactor: create helpers.ts & add Cookie
